LayoutTests/media/video-end.html

<video src=content/test.mp4 end=0.5s controls></video>
<script src=video-test.js></script>
<script>
    testExpected("video.ended", false);
    testExpected("video.end", 0.5);
    waitForEvent('ended', function () {
        testExpected("video.currentTime", 0.5);
        testExpected("video.ended", true);
        endTest();
    } );
    video.play();
</script>
