LayoutTests/media/video-end-expected.txt

EXPECTED (video.ended == 'false') OK
EXPECTED (video.end == '0.5') OK
EVENT(ended)
EXPECTED (video.currentTime == '0.5') OK
EXPECTED (video.ended == 'true') OK
END OF TEST
